+++ docs/changes.rst
@@ -43,6 +43,8 @@ Released: not yet
 * Avoided `DeprecationWarning` on Python 3 for invalid escape sequences
   in some places.
 
+* Added a `__repr__()` function to the `Session` class, for debug purposes.
+
 **Known issues:**
 
 * See `list of open issues`_.